docs(website): Refine playground executor
