+++ website/package.json
@@ -46,6 +46,7 @@
     "react-simple-resizer": "^2.1.0",
     "react-use": "^17.2.3",
     "serialize-error": "^8.1.0",
+    "use-debounce": "^6.0.1",
     "use-immer": "^0.5.1"
   },
   "devDependencies": {
