+++ website/src/modules/playground/components/Sidebar/generateEntries.ts
@@ -1,27 +1,6 @@
 import { groupBy } from "lodash";
 import { sep } from "@site/src/utils/path";
-
-export enum EntryType {
-  File,
-  Directory
-}
-
-interface BaseEntry {
-  type: EntryType;
-  name: string;
-  path: string;
-}
-
-export interface File extends BaseEntry {
-  type: EntryType.File;
-}
-
-export interface Directory extends BaseEntry {
-  type: EntryType.Directory;
-  children: readonly Entry[];
-}
-
-export type Entry = File | Directory;
+import { Entry, EntryType, File, Directory } from "./types";
 
 function doGenerateEntries(
   paths: readonly string[],
