+++ website/src/modules/playground/worker/execute.ts
@@ -1,58 +1,51 @@
 type ExecuteCallback = (err?: unknown, content?: string) => void;
 type Dispose = () => void;
 
-const CATCH_GLOBAL_ERROR = `
-<script type="module">
-  import { serializeError } from "https://cdn.skypack.dev/serialize-error";
+export function execute(
+  id: string,
+  code: string,
+  callback: ExecuteCallback
+): Dispose {
+  function handleError(event: ErrorEvent) {
+    callback(event.error);
+  }
 
-  window.addEventListener("error", (event) => {
-    window.parent.postMessage({
-      source: "kosko-playground",
-      type: "error",
-      payload: serializeError(event.error)
-    });
-  });
-</script>
-`;
-
-export function execute(code: string, callback: ExecuteCallback): Dispose {
-  function handleMessage(event: MessageEvent) {
-    if (event.data.source !== "kosko-playground") {
+  (window as any)[id] = (event) => {
+    if (event.source !== "kosko-playground") {
       return;
     }
 
-    switch (event.data.type) {
+    switch (event.type) {
       case "success":
-        callback(null, event.data.payload);
+        callback(null, event.payload);
         return;
 
       case "error":
-        callback(event.data.payload);
+        callback(event.payload);
         return;
     }
-  }
+  };
 
-  window.addEventListener("message", handleMessage);
+  window.addEventListener("error", handleError);
 
-  const frame = document.createElement("iframe");
-  const blob = new Blob(
-    [CATCH_GLOBAL_ERROR, `<script type="module">${code}</script>`],
-    {
-      type: "text/html"
-    }
-  );
-  const url = URL.createObjectURL(blob);
+  const script = document.createElement("script");
 
-  frame.width = "0";
-  frame.height = "0";
-  frame.src = url;
-  frame.style.display = "none";
+  script.type = "module";
+  script.id = `${id}-script`;
+  script.innerHTML = code;
 
-  document.body.appendChild(frame);
+  script.addEventListener("error", () => {
+    callback({
+      name: "NetworkError",
+      message: "Script load failed. Open the console for more details."
+    });
+  });
+
+  document.body.appendChild(script);
 
   return () => {
-    window.removeEventListener("message", handleMessage);
-    frame.remove();
-    URL.revokeObjectURL(url);
+    delete (window as any)[id];
+    window.removeEventListener("error", handleError);
+    script.remove();
   };
 }
